backend/src/consumers/requirements.txt
flask==2.3.3
flask-socketio==5.1.1
kafka-python==2.0.2
python-socketio==5.4.0
python-dotenv==0.19.0
websocket-client>=1.6.0
requests==2.31.0